+++ CHANGELOG.md
@@ -102,12 +102,13 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Added `EventObserver` for adding event broadcaster capability to JS classes.
 - Added `atomic-checkers.js` and `validateDomElement`
   utility method for checking atomic element DOM nodes.
-- Backend Organisms Full Width Text & Post Preview  
+- Backend Organisms Full Width Text & Post Preview.
 - Added Related Posts molecule to the CFGOVPage
 - Add Main Contact Info molecule
 - Add Sidefoot Streamfield to CFGOVPage for sidebar/footer content
 - Add global context variable `global_dict` for easier prototyping
 - Add styleguide app to local settings
+- Added templates and CSS for the Filterable-List-Controls organism.
 
 ### Changed
 - Updated the primary nav to move focus as user enters and leaves nav levels